tests/wpt/web-platform-tests/orientation-event/device-orientation-events-of-detached-documents.https.html

<!DOCTYPE html>
<title>Device sensor event listeners for `window` of detached documents.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="resources/orientation-event-helpers.js"></script>
<script>

sensor_test(async t => {
  const childFrame = document.createElement('iframe');
  childFrame.src = "/common/blank.html";
  document.body.append(childFrame);

  const childLoadWatcher = new EventWatcher(t, childFrame, ["load"]);
  await childLoadWatcher.wait_for("load");

  const contentWindow = childFrame.contentWindow;
  const contentDocument = childFrame.contentDocument;

  document.body.remove(childFrame);

  assert_not_equals(contentWindow, null);
  assert_not_equals(contentDocument, null);
  assert_equals(contentDocument.defaultView, null);

  contentWindow.addEventListener("devicemotion", () => {});
  contentWindow.addEventListener("deviceorientation", () => {});
  contentWindow.addEventListener("deviceorientationabsolute", () => {});

}, 'Adding an event listener on the window of a detached document does not crash.');

</script>
